+++ react/components/SKUModal/SKUModal.tsx
@@ -20,18 +20,27 @@ import * as yup from 'yup'
 
 import type { SKUType } from '../../context/RegisterContext'
 import { messages } from '../../utils/messages'
+import SKUInput from './SKUInput'
+
+interface FormProps {
+  skuId: string
+  priceTable: string
+  minValue: number
+  maxValue: number
+  defaultValue: number
+}
 
 interface Props {
   handleClose: (form: SKUType) => void
   modalState: ModalStateReturn
-  initialValue?: SKUType
+  initialValue?: FormProps
 }
 
 const SKUModal = (props: Props) => {
   const { handleClose, modalState, initialValue } = props
   const intl = useIntl()
 
-  const defaultValue: SKUType = {
+  const defaultValue: FormProps = {
     skuId: '',
     priceTable: '',
     minValue: 0,
@@ -84,9 +93,12 @@ const SKUModal = (props: Props) => {
       ),
   })
 
-  const handleSubmit = (values: SKUType, actions: FormikHelpers<SKUType>) => {
+  const handleSubmit = (
+    values: FormProps,
+    actions: FormikHelpers<FormProps>
+  ) => {
     const newSKU: SKUType = {
-      skuId: values.skuId,
+      skuId: Number(values.skuId),
       priceTable: values.priceTable,
       minValue: values.minValue,
       maxValue: values.maxValue,
@@ -115,7 +127,7 @@ const SKUModal = (props: Props) => {
         >
           {({ values, isValid, dirty, resetForm }) => (
             <Form>
-              <ModalHeader title="SKU">
+              <ModalHeader title={intl.formatMessage(messages.SKUModalTitle)}>
                 <ModalButton
                   variant="adaptative-dark"
                   closeModalOnClick
@@ -127,10 +139,13 @@ const SKUModal = (props: Props) => {
               </ModalHeader>
               <ModalContent>
                 <Flex direction="column">
-                  <FormikInput name="skuId" label="SKU ID" />
+                  <SKUInput
+                    name="skuId"
+                    label={intl.formatMessage(messages.SKUId)}
+                  />
                   <FormikInput
                     name="priceTable"
-                    label={`${intl.formatMessage(messages.SKUPriceTableLabel)}`}
+                    label={intl.formatMessage(messages.SKUPriceTableLabel)}
                   />
                 </Flex>
                 <Heading csx={{ marginTop: 5 }}>
@@ -165,7 +180,7 @@ const SKUModal = (props: Props) => {
                         <FormikNumericStepper
                           name="minValue"
                           minValue={0}
-                          label={`${intl.formatMessage(messages.SKUItemMin)}`}
+                          label={intl.formatMessage(messages.SKUItemMin)}
                         />
                       </Box>
                       <Box
@@ -190,7 +205,7 @@ const SKUModal = (props: Props) => {
                         <FormikNumericStepper
                           name="maxValue"
                           minValue={values.minValue}
-                          label={`${intl.formatMessage(messages.SKUItemMax)}`}
+                          label={intl.formatMessage(messages.SKUItemMax)}
                         />
                       </Box>
                       <Box
@@ -214,9 +229,7 @@ const SKUModal = (props: Props) => {
                         </Label>
                         <FormikNumericStepper
                           name="defaultValue"
-                          label={`${intl.formatMessage(
-                            messages.SKUItemInitial
-                          )}`}
+                          label={intl.formatMessage(messages.SKUItemInitial)}
                         />
                       </Box>
                       <Box